Move Dashboard components to Project/ directory
