+++ src/components/Project/ProjectActivity/styles.ts
