+++ src/components/Project/modals/ConfirmPayOwnerModal/ProjectRiskNotice.tsx
