+++ src/components/Project/styles.ts
